refactor: Moved strings to constants, and findWebsiteYAML function to utils
